requirements.txt
click==8.1.3
kubernetes==24.2.0
pick==1.4.0
PyYAML==6.0
